fourth pass

+++ NeovaDashboard/Source/UI/SubComponents/HeaderComponent.cpp
@@ -36,10 +36,10 @@ void HeaderComponent::paint (Graphics& g)
     g.setColour (neova_dash::colour::mainText);
 
     g.setFont (neova_dash::font::neovaFont.withHeight (13.5f).withExtraKerningFactor (0.27f));
-    g.drawText ("NEOVA", area.removeFromTop (area.getHeight()/2).withTrimmedBottom (spacing/4), Justification::centredBottom, false);
+    g.drawText ("NEOVA", area.removeFromTop (area.getHeight () / 2).withTrimmedBottom (static_cast<int>(spacing / 4.0f)), Justification::centredBottom, false);
 
   	g.setFont(neova_dash::font::dashFontLight.withHeight (9.5f).withExtraKerningFactor (0.38f));
-  	g.drawText("DASH TESTS", area.withTrimmedTop (spacing*3/4), Justification::centredTop, false);
+    g.drawText ("DASH TESTS", area.withTrimmedTop (static_cast<int>(spacing * 3.0f / 4.0f)), Justification::centredTop, false);
 }
 
 void HeaderComponent::resized()
@@ -317,11 +317,11 @@ void HeaderComponent::BatteryComponent::drawBatteryPath (Graphics& g, juce::Rect
                                                  : lastBattery < 0.4f ? Colours::yellow
                                                  : Colours::lime;
 
-    const int indicatorAreaWidth = area.getWidth()/4;
+    const int indicatorAreaWidth = static_cast<int>(area.getWidth () / 4);
 
     for (int indicator = 0; indicator < 4; indicator++)
     {
-        auto indicatorArea = area.removeFromLeft (indicatorAreaWidth);
+        auto indicatorArea = area.removeFromLeft (static_cast<float>(indicatorAreaWidth));
 
         if (indicator < numIndicators || numBlinkingIndicators == 4)
         {
